refactor: Removed redundant code
