+++ src/components/SettingsPage/index.ts
@@ -1,3 +1,12 @@
-export * from './SettingsPage';
+import { RouteConfig } from '@bluebase/components';
+import { SettingsPageItemProps } from '../SettingsPageItem';
+
+/**
+ * A UI item displayed on a settings details page
+ */
+export interface SettingsPageProps extends RouteConfig {
+	items: SettingsPageItemProps[]
+}
+
 export * from './SettingsPageDesktop';
 export * from './SettingsPageMobile';
